remove Url property (#179)
* remove Url property

* fix url
