+++ models/WikinewsPage.php
@@ -20,8 +20,6 @@ use yii\db\ActiveRecord;
 class WikinewsPage extends ActiveRecord
 {
 	
-	public $url;
-	
     /**
      * {@inheritdoc}
      */
@@ -36,42 +34,10 @@ class WikinewsPage extends ActiveRecord
     public function rules()
     {
         return [
-            [['url'], 'required'],
             [['title'], 'string', 'max' => 255],
             [['language_id', 'group_id', 'pageid', 'created_by', 'created_at', 'parsed_at'], 'integer'],
-            ['url', 'validateUrl'],
         ];
     }
-	
-	/**
-     * Validates the url.
-     * This method serves as the inline validation for url.
-     *
-     * @param string $attribute the attribute currently being validated
-     * @param array $params the additional name-value pairs given in the rule
-     */
-	public function validateUrl($attribute, $params) 
-	{
-		$valid = true;
-		$attr = $this->$attribute;
-		$validateUrl = preg_match("#https://([a-z]{2}).wikinews.org/wiki/[A-Za-z0-9,_.-]+$#", $attr, $matches);
-		if (!$validateUrl) {
-			$valid = false;
-		}
-		elseif ($matches[1]) {
-			$langCode = $matches[1];
-			$langValid = WikinewsLanguage::find()->where(['code'=>$langCode])->count();
-			if(!$langValid) {
-				$valid = false;
-			}
-		}
-		else {
-			$valid = false;
-		}
-		if (!$valid) {
-			$this->addError($attribute, 'Url is not valid.');
-		}
-	}
 
     /**
      * {@inheritdoc}
@@ -82,7 +48,6 @@ class WikinewsPage extends ActiveRecord
             'id' => 'ID',
             'language_id' => 'Language ID',
             'title' => 'Title',
-            'url' => 'Wikinews page url',
             'group_id' => 'Group ID',
             'pageid' => 'Page ID',
             'created_by' => 'Created by',
